Changed the Player and Entity Class for more ease of use.
Changed the Main and Launcher to support arguments being passed through to change player count and screen size.

+++ src/Entity/BaseEntity.java
@@ -34,15 +34,15 @@ public class BaseEntity {
         return z;
     }
     
-    public void setX(int i) {
+    public void setX(float i) {
         x = i;
     }
     
-    public void setY(int i) {
+    public void setY(float i) {
         y = i;
     }
     
-    public void setZ(int i) {
+    public void setZ(float i) {
         z = i;
     }
 }
